refactor the metrics
