+++ plugins/aws/resoto_plugin_aws/utils.py
@@ -1,7 +1,8 @@
 import uuid
 from configparser import ConfigParser
 from pathlib import Path
-from typing import Any, Callable, Dict, Iterable, List, Optional, TypeVar
+from typing import Any, Callable, Dict, Iterable, List, Optional, TypeVar, Tuple
+from attrs import define
 
 from boto3.session import Session as BotoSession
 from botocore.exceptions import ConnectionClosedError, CredentialRetrievalError
@@ -191,3 +192,11 @@ T = TypeVar("T")
 
 def identity(x: T) -> T:
     return x
+
+
+@define(kw_only=True, frozen=True)
+class MetricNormalization:
+    name: Tuple[str, ...]
+    stat_map: Optional[Dict[str, str]] = None
+    unit_map: Optional[Dict[str, str]] = None
+    normalize_value: Callable[[float], float] = identity
